Do away with the vectors completely
A LOT of changes to the project. I completely replaced vectors with C arrays because of the allocation issues.
This has completely fixed the problem I had before and the project compiles.
Note however that it doesn't mean the project actually works. The produced DSiWare export doesn't get accepted by Frogtool, so our work here is not done.